Invoice Matching Project
Goal:   
Compare the PO document with the two invoices provided and the system should reject incorrect invoice and identify all the reasons for incorrectness and should approve the correct invoice

Situation:
Company X has a financial department who releases Purchase Order (PO) to its vendors for procurement of items like hardware, software and services. After the vendor supplies goods, company X receives the goods and the vendor submits a invoice to company X. Currently Company X has a manual process by which a financial processor verifies the invoices by comparing the details of the invoice with the PO and either accepts the invoice and processes payments or rejects the invoice due to errors. Errors could be many type – some of the common ones are either the quantity of item on invoice is not same as PO or price is not same as PO. 
Company X spends considerable amount of manual effort currently as it has a high volume of POs and Invoices running into lakhs of invoices per month and it has a team of 10 financial processors doing manual verification. Company X wants to build a Next Generation PO Invoice Matching system which eliminate the manual process and automate the whole verification process. 
Challenge:
Please design a web application in the language of your choice which will be given to Company X’s financial processors and it should automate the verification process and say if the invoice is matching and accept it or reject it. Please view the sample PO and sample invoices for reference.  You can use any language that you are comfortable with and output can be viewed on any device (either web or mobile or both ).